+++ src/components/share/modal/items/ShareLinkSettings.tsx
@@ -2,9 +2,14 @@ import { DropdownMenu } from ":/components/dropdown-menu/DropdownMenu";
 import { useDropdownMenu } from ":/components/dropdown-menu/useDropdownMenu";
 import LogoPublic from ":/components/share/assets/public.svg";
 import LogoPrivate from ":/components/share/assets/lock_person.svg";
-import { Button, useCunningham } from "@openfun/cunningham-react";
+import { Button } from "@openfun/cunningham-react";
 import { useMemo, useState } from "react";
 import { DropdownMenuOption } from ":/components/dropdown-menu";
+import {
+  CustomTranslations,
+  useCustomTranslations,
+} from ":/hooks/useCustomTranslations";
+import { TranslationKey } from ":/types/translations";
 
 const getIcon = (value: string | undefined) => {
   switch (value) {
@@ -26,6 +31,7 @@ export const ShareLinkSettings = ({
   linkRoleChoices = [],
   linkRole = "reader",
   onUpdateLinkRole,
+  customTranslations,
 }: {
   canUpdate?: boolean;
   linkReachChoices?: Partial<DropdownMenuOption>[];
@@ -35,6 +41,7 @@ export const ShareLinkSettings = ({
   linkRoleChoices?: Partial<DropdownMenuOption>[];
   linkRole?: "reader" | "editor";
   onUpdateLinkRole: (value: string) => void;
+  customTranslations?: CustomTranslations;
 }) => {
   const linkReachDropdown = useDropdownMenu();
   const linkRoleDropdown = useDropdownMenu();
@@ -46,7 +53,7 @@ export const ShareLinkSettings = ({
     string[]
   >([linkRole]);
   const selectedLinkRole = selectedLinkRoleValues[0];
-  const { t } = useCunningham();
+  const { t } = useCustomTranslations(customTranslations);
 
   const reachChoices = useMemo(
     () =>
@@ -57,7 +64,7 @@ export const ShareLinkSettings = ({
           label: t(
             "components.share.linkSettings.reach.choices." +
               choice.value +
-              ".title"
+              ".title" as TranslationKey
           ),
           ...choice,
         };
@@ -78,7 +85,7 @@ export const ShareLinkSettings = ({
           label: t(
             "components.share.linkSettings.role.choices." +
               choice.value +
-              ".title"
+              ".title" as TranslationKey
           ),
           ...choice,
         };
@@ -189,14 +196,14 @@ export const ShareLinkSettings = ({
         {renderLinkReach()}
         <div className="c__share-modal__link-settings__content__description desktop">
           {t(
-            `components.share.linkSettings.reach.choices.${selectedLinkReach}.description`
+            `components.share.linkSettings.reach.choices.${selectedLinkReach}.description` as TranslationKey
           )}
         </div>
         {renderLinkRole()}
       </div>
       <div className="c__share-modal__link-settings__content__description mobile">
         {t(
-          `components.share.linkSettings.reach.choices.${selectedLinkReach}.description`
+          `components.share.linkSettings.reach.choices.${selectedLinkReach}.description` as TranslationKey
         )}
       </div>
     </div>